data-generator/requirements.txt
kafka-python==2.0.2
faker==19.6.2 